backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# MongoDB
motor==3.3.2
pymongo==4.6.0

# Vector Databases
pymilvus==2.3.4
faiss-cpu==1.7.4

# ML/AI Models
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
sentence-transformers==2.2.2
openai-clip==1.0.1
openai==1.3.7
Pillow==10.1.0

# Audio Processing
openai-whisper==20231117
soundfile==0.12.1
pydub==0.25.1

# Geospatial
geopy==2.4.1
shapely==2.0.2
pyproj==3.6.1

# HTTP Clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
imagehash==4.3.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
